+++ src/main/java/com/atomgraph/linkeddatahub/server/filter/request/OntologyFilter.java
@@ -19,7 +19,7 @@ package com.atomgraph.linkeddatahub.server.filter.request;
 import com.atomgraph.linkeddatahub.apps.model.Application;
 import com.atomgraph.linkeddatahub.apps.model.EndUserApplication;
 import com.atomgraph.linkeddatahub.vocabulary.LAPP;
-import com.atomgraph.processor.exception.OntologyException;
+import com.atomgraph.server.exception.OntologyException;
 import com.atomgraph.linkeddatahub.server.util.OntologyModelGetter;
 import java.io.IOException;
 import java.net.URI;